refactor: 바이너리 컨텐츠 코어 로직 유저, 메세지 도메인으로 이동
s3client가 아니라 s3Adapter를 모킹하면 SUCCESS가 아니라 Waiting이 나옵니다. 비동기 처리시 모킹 주의
