+++ src/main/java/com/sprint/mission/discodeit/domain/user/service/basic/UserBinaryContentService.java
@@ -1,13 +1,10 @@
-package com.sprint.mission.discodeit.domain.binarycontent.service.basic;
+package com.sprint.mission.discodeit.domain.user.service.basic;
 
 import com.sprint.mission.discodeit.domain.binarycontent.dto.BinaryContentRequest;
-import com.sprint.mission.discodeit.domain.binarycontent.dto.BinaryContentResult;
 import com.sprint.mission.discodeit.domain.binarycontent.entity.BinaryContent;
 import com.sprint.mission.discodeit.domain.binarycontent.exception.BinaryContentNotFoundException;
 import com.sprint.mission.discodeit.domain.binarycontent.repository.BinaryContentRepository;
-import com.sprint.mission.discodeit.domain.binarycontent.service.BinaryContentService;
 import com.sprint.mission.discodeit.domain.binarycontent.storage.BinaryContentStorage;
-import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -16,17 +13,15 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
-
 @Service
 @RequiredArgsConstructor
-public class BasicBinaryContentService implements BinaryContentService {
+public class UserBinaryContentService {
 
-  private final BinaryContentStorage binaryContentStorage;
   private final BinaryContentRepository binaryContentRepository;
+  private final BinaryContentStorage binaryContentStorage;
 
   @Transactional
-  @Override
-  public BinaryContentResult createBinaryContent(BinaryContentRequest binaryContentRequest) {
+  public BinaryContent createBinaryContent(BinaryContentRequest binaryContentRequest) {
     if (binaryContentRequest == null) {
       return null;
     }
@@ -34,34 +29,25 @@ public class BasicBinaryContentService implements BinaryContentService {
         binaryContentRequest.contentType(), binaryContentRequest.size());
     BinaryContent savedBinaryContent = binaryContentRepository.save(binaryContent);
 
-    binaryContentStorage.put(savedBinaryContent.getId(), binaryContentRequest.bytes());
+    TransactionSynchronizationManager.registerSynchronization(
+        new TransactionSynchronization() {
+          @Override
+          public void afterCommit() {
+            binaryContentStorage.put(savedBinaryContent.getId(), binaryContentRequest.bytes());
+          }
+        }
+    );
 
-    return BinaryContentResult.fromEntity(savedBinaryContent);
+    return savedBinaryContent;
   }
 
-  @Transactional(readOnly = true)
-  @Override
-  public BinaryContentResult getById(UUID id) {
-    BinaryContent binaryContent = binaryContentRepository.findById(id)
-        .orElseThrow(() -> new BinaryContentNotFoundException(Map.of()));
+  public void delete(BinaryContent binaryContent) {
+    if (binaryContent == null) {
+      return;
+    }
+    validateBinaryContentExist(binaryContent.getId());
 
-    return BinaryContentResult.fromEntity(binaryContent);
-  }
-
-  @Transactional(readOnly = true)
-  @Override
-  public List<BinaryContentResult> getByIdIn(List<UUID> ids) {
-    return binaryContentRepository.findAllById(ids)
-        .stream()
-        .map(BinaryContentResult::fromEntity)
-        .toList();
-  }
-
-  @Override
-  public void delete(UUID id) {
-    validateBinaryContentExist(id);
-
-    binaryContentRepository.deleteById(id);
+    binaryContentRepository.deleteById(binaryContent.getId());
   }
 
   private void validateBinaryContentExist(UUID id) {
