+++ src/test/java/com/sprint/mission/discodeit/domain/binarycontent/service/BasicBinaryContentServiceTest.java
@@ -1,18 +1,24 @@
 package com.sprint.mission.discodeit.domain.binarycontent.service;
 
-import com.sprint.mission.discodeit.testutil.IntegrationTestSupport;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
+import com.sprint.mission.discodeit.domain.binarycontent.dto.BinaryContentRequest;
 import com.sprint.mission.discodeit.domain.binarycontent.dto.BinaryContentResult;
 import com.sprint.mission.discodeit.domain.binarycontent.entity.BinaryContent;
 import com.sprint.mission.discodeit.domain.binarycontent.exception.BinaryContentNotFoundException;
 import com.sprint.mission.discodeit.domain.binarycontent.repository.BinaryContentRepository;
+import com.sprint.mission.discodeit.domain.binarycontent.storage.BinaryContentStorage;
+import com.sprint.mission.discodeit.testutil.IntegrationTestSupport;
+import java.util.List;
+import java.util.UUID;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-
-import java.util.List;
-import java.util.UUID;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 class BasicBinaryContentServiceTest extends IntegrationTestSupport {
 
@@ -21,10 +27,42 @@ class BasicBinaryContentServiceTest extends IntegrationTestSupport {
   @Autowired
   private BinaryContentService binaryContentService;
 
-  @AfterEach
-  void tearDown() {
+  @MockitoBean
+  private BinaryContentStorage binaryContentStorage;
+
+  @BeforeEach
+  void beforeEach() {
     binaryContentRepository.deleteAllInBatch();
   }
+
+  @DisplayName("바이너리 컨텐츠를 생성합니다.")
+  @Test
+  void createBinaryContent() {
+    // given
+    String name = UUID.randomUUID().toString();
+    BinaryContentRequest binaryContentRequest = new BinaryContentRequest(name, "", 0,
+        "hello".getBytes());
+
+    // when
+    BinaryContentResult binaryContent = binaryContentService.createBinaryContent(
+        binaryContentRequest);
+
+    // then
+    Assertions.assertThat(binaryContent.name()).isEqualTo(name);
+
+    verify(binaryContentStorage, times(1)).put(any(), any());
+  }
+
+  @DisplayName("null 받을 경우, null을 리턴합니다.")
+  @Test
+  void createBinaryContent_Null() {
+    // when
+    BinaryContentResult binaryContent = binaryContentService.createBinaryContent(null);
+
+    // then
+    Assertions.assertThat(binaryContent).isNull();
+  }
+
 
   @DisplayName("ID로 조회하면, 해당 객체를 반환한다.")
   @Test
@@ -62,6 +100,27 @@ class BasicBinaryContentServiceTest extends IntegrationTestSupport {
     Assertions.assertThat(binaryContentResults)
         .extracting(BinaryContentResult::id)
         .containsExactlyInAnyOrder(firstBinaryContent.getId(), secondBinaryContent.getId());
+  }
+
+  @DisplayName("데이터에 저장된 바이너리 컨텐츠를 삭제합니다.")
+  @Test
+  void delete() {
+    // given
+    BinaryContent binaryContent = binaryContentRepository.save(new BinaryContent("", "", 10));
+
+    // when
+    binaryContentService.delete(binaryContent.getId());
+
+    // then
+    Assertions.assertThat(binaryContentRepository.findById(binaryContent.getId())).isNotPresent();
+  }
+
+  @DisplayName("바이너리 컨텐츠를 삭제시, 해당 객체가 없으면 삭제합니다.")
+  @Test
+  void delete_NoException() {
+    // when & then
+    Assertions.assertThatThrownBy(() -> binaryContentService.delete(UUID.randomUUID()))
+        .isInstanceOf(BinaryContentNotFoundException.class);
   }
 
 }
